Close #515 - Reorganize tests
