+++ modules/effectie-cats-effect2/shared/src/test/scala-2/effectie/instances/ce2/canCatchSpec.scala
@@ -89,7 +89,7 @@ object canCatchSpec extends Properties {
   )
 
   /* Future */
-  val futureSpecs = effectie.core.CanCatchSpec.futureSpecs ++ List(
+  val futureSpecs = effectie.instances.future.canCatchSpec.futureSpecs ++ List(
     example(
       "test CanCatch[Future]catchNonFatalEitherT should catch NonFatal",
       FutureSpec.testCanCatch_Future_catchNonFatalEitherTShouldCatchNonFatal,
