+++ modules/effectie-cats-effect2/shared/src/test/scala-2/effectie/instances/ce2/canHandleErrorSpec.scala
@@ -152,7 +152,7 @@ object canHandleErrorSpec extends Properties {
   )
 
   /* Future */
-  private val futureSpecs = effectie.core.CanHandleErrorSpec.futureSpecs ++ List(
+  private val futureSpecs = effectie.instances.future.canHandleErrorSpec.futureSpecs ++ List(
     example(
       "test CanHandleError[Future].handleEitherTNonFatalWith should handle NonFatal",
       FutureSpec.testCanHandleError_Future_handleEitherTNonFatalWithShouldHandleNonFatalWith,
