+++ modules/effectie-cats-effect2/shared/src/test/scala-2/effectie/instances/ce2/canRecoverSpec.scala
@@ -154,7 +154,7 @@ object canRecoverSpec extends Properties {
   )
 
   /* Future */
-  val futureSpecs = effectie.core.CanRecoverSpec.futureSpecs ++ List(
+  val futureSpecs = effectie.instances.future.canRecoverSpec.futureSpecs ++ List(
     example(
       "test CanRecover[Future].recoverEitherTFromNonFatalWith should catch NonFatal",
       FutureSpec.testCanRecover_Future_recoverEitherTFromNonFatalWithShouldRecoverFromNonFatal,
