+++ modules/effectie-cats-effect2/shared/src/test/scala-2/effectie/instances/ce2/fxCtorSpec.scala
@@ -75,7 +75,7 @@ object fxCtorSpec extends Properties {
     ),
   )
 
-  private val futureSpecs = effectie.core.FxCtorSpec.futureSpecs
+  private val futureSpecs = effectie.instances.future.fxCtorSpec.futureSpecs
 
   private val idSpecs = {
     import effectie.instances.id.fxCtor._
