+++ modules/effectie-cats-effect2/shared/src/test/scala-2/effectie/instances/ce2/fxSpec.scala
@@ -368,7 +368,7 @@ object fxSpec extends Properties {
     )
 
   /* Future */
-  private val futureSpecs = effectie.core.FxSpec.futureSpecs ++
+  private val futureSpecs = effectie.instances.future.fxSpec.futureSpecs ++
     List(
       example(
         "test Fx[Future]catchNonFatalEitherT should catch NonFatal",
