+++ modules/effectie-cats-effect2/shared/src/test/scala-3/effectie/instances/ce2/canCatchSpec.scala
@@ -87,7 +87,7 @@ object CanCatchSpec extends Properties {
   )
 
   /* Future */
-  val futureSpecs = effectie.core.CanCatchSpec.futureSpecs ++ List(
+  val futureSpecs = effectie.instances.future.canCatchSpec.futureSpecs ++ List(
     example(
       "test CanCatch[Future]catchNonFatalEitherT should catch NonFatal",
       FutureSpec.testCanCatch_Future_catchNonFatalEitherTShouldCatchNonFatal,
