+++ modules/effectie-cats-effect2/shared/src/test/scala-3/effectie/instances/ce2/fxCtorSpec.scala
@@ -71,7 +71,7 @@ object FxCtorSpec extends Properties {
     ),
   )
 
-  private val futureSpecs = effectie.core.FxCtorSpec.futureSpecs
+  private val futureSpecs = effectie.instances.future.fxCtorSpec.futureSpecs
 
   private val idSpecs = {
     import effectie.instances.id.fxCtor.*
