+++ modules/effectie-cats-effect2/shared/src/test/scala-3/effectie/instances/ce2/fxSpec.scala
@@ -365,7 +365,7 @@ object FxSpec extends Properties {
     )
 
   /* Future */
-  private val futureSpecs = effectie.core.FxSpec.futureSpecs ++
+  private val futureSpecs = effectie.instances.future.fxSpec.futureSpecs ++
     List(
       example(
         "test Fx[Future]catchNonFatalEitherT should catch NonFatal",
