+++ modules/effectie-cats-effect3/shared/src/test/scala-2/effectie/instances/ce3/canCatchSpec.scala
@@ -91,7 +91,7 @@ object canCatchSpec extends Properties {
   )
 
   /* Future */
-  val futureSpecs = effectie.core.CanCatchSpec.futureSpecs ++ List(
+  val futureSpecs = effectie.instances.future.canCatchSpec.futureSpecs ++ List(
     example(
       "test CanCatch[Future]catchNonFatalEitherT should catch NonFatal",
       FutureSpec.testCanCatch_Future_catchNonFatalEitherTShouldCatchNonFatal,
