+++ modules/effectie-cats-effect3/shared/src/test/scala-2/effectie/instances/ce3/canHandleErrorSpec.scala
@@ -157,7 +157,7 @@ object canHandleErrorSpec extends Properties {
     )
 
   /* Future */
-  private val futureSpecs = effectie.core.CanHandleErrorSpec.futureSpecs ++ List(
+  private val futureSpecs = effectie.instances.future.canHandleErrorSpec.futureSpecs ++ List(
     example(
       "test CanHandleError[Future].handleEitherTNonFatalWith should handle NonFatal",
       FutureSpec.testCanHandleError_Future_handleEitherTNonFatalWithShouldHandleNonFatalWith,
