+++ modules/effectie-cats-effect3/shared/src/test/scala-2/effectie/instances/ce3/fxCtorSpec.scala
@@ -108,7 +108,7 @@ object fxCtorSpec extends Properties {
     ),
   )
 
-  private val futureSpecs = effectie.core.FxCtorSpec.futureSpecs
+  private val futureSpecs = effectie.instances.future.fxCtorSpec.futureSpecs
 
   private val idSpecs = {
     import effectie.instances.id.fxCtor._
