+++ modules/effectie-cats-effect3/shared/src/test/scala-2/effectie/instances/ce3/fxSpec.scala
@@ -416,7 +416,7 @@ object fxSpec extends Properties {
     )
 
   /* Future */
-  private val futureSpecs = effectie.core.FxSpec.futureSpecs ++
+  private val futureSpecs = effectie.instances.future.fxSpec.futureSpecs ++
     List(
       example(
         "test Fx[Future]catchNonFatalEitherT should catch NonFatal",
