+++ modules/effectie-cats-effect3/shared/src/test/scala-3/effectie/instances/ce3/canCatchSpec.scala
@@ -90,7 +90,7 @@ object canCatchSpec extends Properties {
   )
 
   /* Future */
-  val futureSpecs = effectie.core.CanCatchSpec.futureSpecs ++ List(
+  val futureSpecs = effectie.instances.future.canCatchSpec.futureSpecs ++ List(
     example(
       "test CanCatch[Future]catchNonFatalEitherT should catch NonFatal",
       FutureSpec.testCanCatch_Future_catchNonFatalEitherTShouldCatchNonFatal,
