+++ modules/effectie-cats-effect3/shared/src/test/scala-3/effectie/instances/ce3/fxSpec.scala
@@ -421,7 +421,7 @@ object FxSpec extends Properties {
     )
 
   /* Future */
-  private val futureSpecs = effectie.core.FxSpec.futureSpecs ++
+  private val futureSpecs = effectie.instances.future.fxSpec.futureSpecs ++
     List(
       example(
         "test Fx[Future]catchNonFatalEitherT should catch NonFatal",
