+++ modules/effectie-cats/shared/src/test/scala/effectie/instances/id/canCatchSpec.scala
@@ -27,7 +27,7 @@ object canCatchSpec extends Properties {
   override def tests: List[Test] = futureSpecs ++ idSpecs
 
   /* Future */
-  val futureSpecs = effectie.core.CanCatchSpec.futureSpecs ++ List(
+  val futureSpecs = effectie.instances.future.canCatchSpec.futureSpecs ++ List(
     example(
       "test CanCatch[Future]catchNonFatalEitherT should catch NonFatal",
       FutureSpec.testCanCatch_Future_catchNonFatalEitherTShouldCatchNonFatal,
