+++ modules/effectie-cats/shared/src/test/scala/effectie/instances/id/canHandleErrorSpec.scala
@@ -26,7 +26,7 @@ object canHandleErrorSpec extends Properties {
   override def tests: List[Test] = futureSpecs ++ idSpecs
 
   /* Future */
-  private val futureSpecs = effectie.core.CanHandleErrorSpec.futureSpecs ++ List(
+  private val futureSpecs = effectie.instances.future.canHandleErrorSpec.futureSpecs ++ List(
     example(
       "test CanHandleError[Future].handleEitherTNonFatalWith should handle NonFatal",
       FutureSpec.testCanHandleError_Future_handleEitherTNonFatalWithShouldHandleNonFatalWith,
