+++ modules/effectie-cats/shared/src/test/scala/effectie/instances/id/canRecoverSpec.scala
@@ -27,7 +27,7 @@ object canRecoverSpec extends Properties {
       idSpecs
 
   /* Future */
-  val futureSpecs = effectie.core.CanRecoverSpec.futureSpecs ++ List(
+  val futureSpecs = effectie.instances.future.canRecoverSpec.futureSpecs ++ List(
     example(
       "test CanRecover[Future].recoverEitherTFromNonFatalWith should catch NonFatal",
       FutureSpec.testCanRecover_Future_recoverEitherTFromNonFatalWithShouldRecoverFromNonFatal,
