+++ modules/effectie-cats/shared/src/test/scala/effectie/instances/id/fxSpec.scala
@@ -29,7 +29,7 @@ object fxSpec extends Properties {
   private val unit: Unit = ()
 
   /* Future */
-  private val futureSpecs = effectie.core.FxSpec.futureSpecs ++
+  private val futureSpecs = effectie.instances.future.fxSpec.futureSpecs ++
     List(
       example(
         "test Fx[Future]catchNonFatalEitherT should catch NonFatal",
