+++ modules/effectie-core/shared/src/test/scala/effectie/instances/future/canCatchSpec.scala
@@ -1,6 +1,7 @@
-package effectie.core
+package effectie.instances.future
 
 import cats.Functor
+import effectie.core.{CanCatch, FxCtor}
 import effectie.testing.types.SomeError
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
@@ -10,7 +11,7 @@ import hedgehog.runner._
 /** @author Kevin Lee
   * @since 2022-01-05
   */
-object CanCatchSpec extends Properties {
+object canCatchSpec extends Properties {
   private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
   override def tests: List[Test] = futureSpecs
