+++ modules/effectie-core/shared/src/test/scala/effectie/specs/MonadSpec.scala
