+++ modules/effectie-monix3/shared/src/test/scala/effectie/instances/monix3/canHandleErrorSpec.scala
@@ -277,7 +277,7 @@ object canHandleErrorSpec extends Properties {
   )
 
   /* Future */
-  private val futureSpecs = effectie.core.CanHandleErrorSpec.futureSpecs ++ List(
+  private val futureSpecs = effectie.instances.future.canHandleErrorSpec.futureSpecs ++ List(
     example(
       "test CanHandleError[Future].handleEitherTNonFatalWith should handle NonFatal",
       FutureSpec.testCanHandleError_Future_handleEitherTNonFatalWithShouldHandleNonFatalWith,
